refactor(pega_io): move Action Analysis readers into pega_io; add scan_parquet_path
Move read_nested_zip_files, read_gzipped_data, and read_gzipped_ndjson_directory
from decision_analyzer/data_read_utils.py into a new pega_io/action_analysis.py
so every user-facing path → polars read flows through pega_io. Add a thin
scan_parquet_path helper for the folder/glob parquet scans that explanations
needs. Re-export the Action Analysis readers from the old data_read_utils
module (silent shim) and from pega_io/__init__ so callers have one canonical
import path. Update the DA Streamlit upload handler to import from pega_io.
