+++ python/pdstools/pega_io/File.py
@@ -167,6 +167,32 @@ def _read_from_bytesio(file: BytesIO, extension: str) -> pl.LazyFrame:
     raise ValueError(f"Unsupported file type for BytesIO: {extension}")
 
 
+def scan_parquet_path(source: str | Path | list[str] | list[Path]) -> pl.LazyFrame:
+    """Scan one or more parquet files from a user-supplied path, glob, or list.
+
+    Thin wrapper around :func:`polars.scan_parquet` that lives in ``pega_io`` so
+    all user-facing parquet scans flow through a single, inspectable surface.
+    This is the funnel CodeQL's ``py/path-injection`` rule is scoped to.
+
+    Parameters
+    ----------
+    source : str or Path or list of str/Path
+        Path to a parquet file, a glob pattern (e.g. ``"folder/*_BATCH_*.parquet"``),
+        or a list of file paths. Forwarded to ``pl.scan_parquet`` unchanged.
+
+    Returns
+    -------
+    pl.LazyFrame
+        LazyFrame over the matched parquet file(s).
+
+    """
+    # lgtm [py/path-injection]
+    # CodeQL suppression: User-controlled paths are expected in a data reading
+    # library. Users explicitly specify which files/globs to scan — this is the
+    # intended functionality, not a vulnerability.
+    return pl.scan_parquet(source)
+
+
 def read_data(path: str | Path | BytesIO) -> pl.LazyFrame:
     """Read data from various file formats and sources.
