+++ python/pdstools/pega_io/__init__.py
@@ -1,5 +1,10 @@
 from __future__ import annotations
 
+from .action_analysis import (
+    read_gzipped_data,
+    read_gzipped_ndjson_directory,
+    read_nested_zip_files,
+)
 from .Anonymization import Anonymization
 from .API import _read_client_credential_file as _read_client_credential_file
 from .API import get_token
@@ -12,6 +17,7 @@ from .File import (
     read_ds_export,
     read_multi_zip,
     read_zipped_file,
+    scan_parquet_path,
 )
 from .S3 import S3Data
 
@@ -25,6 +31,10 @@ __all__ = [
     "read_data",
     "read_dataflow_output",
     "read_ds_export",
+    "read_gzipped_data",
+    "read_gzipped_ndjson_directory",
     "read_multi_zip",
+    "read_nested_zip_files",
     "read_zipped_file",
+    "scan_parquet_path",
 ]
